Génération
Partie génération de l'algorithme génétique
